fix(build): silence in-tree compiler warnings — sign-conv + mismatched-new-delete
Clean build warning count 373 → 28 (잔존 28건은 모두 외부 Eigen/pinocchio 헤더).

- rtc_base/threading/thread_utils.hpp:121 — vector indexing 에 `static_cast<size_t>(slot)` 추가. 범위 가드는 상위 라인에 이미 존재.
- rtc_mpc test_mpc_handler_alloc_tracer.cpp — GCC 13+ `-Wmismatched-new-delete` 가 global op new/delete + malloc/free 짝을 오진단. `#pragma GCC diagnostic` push/pop 으로 해당 블록만 억제.
- shape_estimation / rtc_mpc test 파일들 — `static_cast<uint64_t>(x) * NNN'ULL` 패턴이 x86_64 Linux 에서 uint64_t (==unsigned long) → unsigned long long rank promotion 으로 sign-conv 발화. `uint64_t{NNN}` brace init 으로 양 피연산자 정렬.

+++ rtc_base/include/rtc_base/threading/thread_utils.hpp
@@ -118,7 +118,7 @@ inline std::string ValidateThreadConfig(const ThreadConfig& cfg) noexcept {
   if (topology.physical_core_slots.empty() ||
       slot >= static_cast<int>(topology.physical_core_slots.size()))
     return slot;
-  return topology.physical_core_slots[slot];
+  return topology.physical_core_slots[static_cast<size_t>(slot)];
 }
 
 [[nodiscard]] inline int SlotToLogicalCpu(int slot) noexcept {
